load-minimizing objective function
I write the model basing on paulina's code. But i do not understand what u(i) here. it seems like fij in the function (26).

But it did not work out so i have to add the third variables fij.

But the model can not find a solution.

I don't know how to continue here and i didn't find some examples regarding this problem.